refactor: decompose search Container; create custom hook
